Add MeriseAcide, a tool to help in DB modeling
You give it a conceptual model (called entity relation model or
something like that in English). Somewhat part of the French method
called Merise, which is (used to?) be taught in France.

The important part of this method is how it makes it comfortable to work
at the conceptual model. Which can then be converted to the physical
model mechanically. In my experience, only at school do you attempt to
formalize relational database modeling. At work, you might do a bit of
it for the prototype of a product, and then you keep adding and patching
as reality catches up.

I wrote this tool at school in 2006 and published in 2007 in France. It
was used by a grand total of maybe three person over the years after.